requirements.txt
temporalio>=1.4.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
langgraph>=0.0.10
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest
